fix(fix(bug):no more warnings): F_Undo + v_Triggerstring
v_Triggerstring for no reason was used as global variable within function F_CreateHotstring. In the body of this function the v_Triggerstring could be local and this was now implemented. As as result no longer warning is triggered on time of application start if undo shortcut is pressed.

Additionally F_InitiateInputHook now initiates some global variables and this function was shifted to "Critical" (protected) section of the script.

Tool: vsc
Stable.