fix(engine): Closing of GUI "Triggerstring tips and hotstring menu styling" blocked the application
It used to be a case that clicking outside the "Triggerstring tips and hotstring menu styling" GUI blocked the whole application and it required restart. This is no longer the case. Also no pressing Esc or just clicking on X button in top right corner just closes the application.

tool: vsc
Stable